src/requirements.txt
flask==2.0.1
werkzeug==2.0.1
requests==2.31.0
gunicorn==20.1.0
pyyaml==6.0